:art: 优化代码结构, 拆分代码
